[Fix]: the 0x02-functions_nested_loops directory files

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,15 +8,15 @@
 int main(void)
 {
 	int c;
-	float ch = 1, k = 2, r = 0;
+	int ch = 1, k = 2, r = 0;
 
 	printf("1, 2");
-	for(c = 1; c <= 96; c++)
+	for (c = 1; c <= 96; c++)
 	{
 		r = ch + k;
 		ch = k;
 		k = r;
-		printf(", %.0f",r);
+		printf(", %.0f", r);
 	}
 	putchar(10);
 	return (0);
